projects/matter-simulation/backend/requirements.txt
flask
flask-cors
flask-session
boto3
pillow
numpy
torch
transformers
scikit-learn
python-dotenv